backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
chromadb==0.4.22
google-generativeai==0.3.2
beautifulsoup4==4.12.3
python-dotenv==1.0.0
lxml==5.1.0
sentence-transformers==2.3.1
